moved AwsTestKekManager to the Abstract class
